breakfast

+++ database/migrations/2025_06_02_044651_add_few_column_to_users_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::table('users', function (Blueprint $table) {
             $table->integer('current_level')->default(1);
-            $table->decimal('current_point', 6,2)->default(0);
+            $table->integer('current_point')->default(0);
             $table->string('nim')->nullable();
             $table->boolean('is_member')->nullable();
             $table->boolean('is_lecturer')->default(false);
